Refactored and removed print lines from expression tree related classes. Add object constructors for Employee and Dependent classes under the object creator class. Also added a method to the ClassContainer to retrieve individual records for constructing objects that need references to existing classes. Eg. Dependent needs to reference an employee.

+++ src/database/Classes/DatabaseClass.java
@@ -6,7 +6,7 @@ import database.utilities.IdGenerator;
  * Created by Kevin on 4/15/2017.
  */
 public class DatabaseClass {
-    private long oid;
+    private int oid;
 
     public DatabaseClass() {
         setOID();
@@ -16,7 +16,7 @@ public class DatabaseClass {
         oid = IdGenerator.getID();
     }
 
-    public long getOID() {
+    public int getOID() {
         return oid;
     }
